pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hh-data-pipeline"
version = "1.0.0"
description = "Пайплайн обработки данных HH с использованием цепочки ответственности"
authors = [
    {name = "Ваше Имя", email = "ваш.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0"
]

[project.scripts]
hh-pipeline = "app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py38']